arrow representation for Tracer
This representation allows for better short-circuiting/laziness. Tracers
which are known to produce no tracing effects can be ignored, allowing
us to boast that often, tracing can be made zero-cost by substituting
the nullTracer appropriately, as opposed to say using CPP to eliminate
tracing in certain builds.

A consequence of this new design is that you cannot use side-effects
when computing which tracer to use. That's a good thing; you really
shouldn't ever have done that to begin with.