allow for side-effects in arrow computation
The goal is to support existing use cases in which benign IOs like
myThreadId and getCurrentTime are used to produce inputs to the tracer
effects that are emitted. Those are now allowed in the arrow composition
by way of category/arrow/arrowchoice, but not monad, so we retain the
ability to see whether a computation will definitely not do any tracing
effects.